Refactor : filter page params -> query 로 변경
